+++ wayland-client/src/lib.rs
@@ -98,7 +98,7 @@ pub mod egl;
 pub mod cursor;
 
 pub use display::{connect_to, default_connect, ConnectError, FatalError};
-pub use env::{EnvHandler, EnvHandlerInner};
+pub use env::{EnvHandler, EnvHandlerInner, EnvNotify};
 pub use event_queue::{EventQueue, EventQueueHandle, ReadEventsGuard, RegisterStatus, State, StateToken};
 
 /// Common routines for wayland proxy objects.